Fix navigate to new transaction bug
Clicking the (+) icon for a new transaction multiple times navigated to
the same fragment multiple times. When this happens, clicking the cancel
button doesn't take you back to the previous fragment. Instead, it took
you back to the most recent new transaction. Fix this by only changing
to new transaction fragment once if not currently on that fragment.