Improve delete procedure
Now that we can select characters with a custom adapter, we can improve
our delete functionality so that we're not deleting by an encoded
Character NamePart anymore. This fixes a bug where a if a character or
campaign had a hyphen in them, we'll be able to still delete the correct
character.