Fix all items selection for all characters bug
When creating a new transaction, the dropdown for the selectable items
were from every character that exists in the DB. We only want to
dropdown items that existed for the current character because not all
campaigns share the same items. This also reduces the list of items
selectable for dropdown.