Fix possible integer overflow bug
It's possible for a integer overflow to happen by creating many
transactions that are huggeeee. This update places reasonable limits and
prefers to use BigDecimal() for preventing overflows.

---
The use of BigDecimal isn't great to use everywhere for all the
calculations due to some reasonable limits, however, in a future update,
we'll allow fractional quantities for items, and to do the math
correctly on those, we can't use double/float types. This just lays the
groundwork to allow fractional quantities as well as make overflow less
of a worry.